move raw_monitor and transect_monitor contents to operations_watchdog
