+++ src/echodataflow/operations/operations_watchdog.py
@@ -2,7 +2,7 @@ from pathlib import Path
 
 from prefect.events import emit_event
 
-from echodataflow.utils.file_watcher import watch_directory
+from echodataflow.utils.file_watcher import watch_directory, watch_file
 from echodataflow.utils.processing_ledger import (
     initialize_ledger,
     register_raw_file,
@@ -11,6 +11,10 @@ from echodataflow.utils.processing_ledger import (
 
 RAW_UPDATE_EVENT = "echodataflow.raw.updated"
 RAW_RESOURCE_ID = "raw-monitor"
+
+TRANSECT_UPDATE_EVENT = "echodataflow.transect.updated"
+TRANSECT_RESOURCE_ID = "transect-start-end-time"
+TRANSECT_RELATED_RESOURCE_ID = "transect-monitor"
 
 
 def emit_raw_update_event(path: Path) -> None:
@@ -55,4 +59,32 @@ def watch_raw_directory(
         directory=raw_directory,
         callback=lambda raw_path: register_and_emit_raw_update(raw_path, db_path),
         pattern="*.raw",
+    )
+
+
+def emit_transect_update_event(path: Path) -> None:
+    """Emit a Prefect event when the transect CSV is updated."""
+
+    emit_event(
+        event=TRANSECT_UPDATE_EVENT,
+        resource={
+            "prefect.resource.id": TRANSECT_RESOURCE_ID,
+            "path": str(path),
+        },
+        related=[
+            {
+                "prefect.resource.id": TRANSECT_RELATED_RESOURCE_ID,
+                "prefect.resource.name": TRANSECT_RELATED_RESOURCE_ID,
+                "prefect.resource.role": "deployment",
+            }
+        ],
+    )
+
+
+def watch_transect_file(path: str | Path):
+    """Watch the transect start/end CSV and emit a Prefect event on update."""
+
+    return watch_file(
+        target_file=path,
+        callback=emit_transect_update_event,
     )
